+++ docs/nuxt.config.ts
@@ -1,3 +1,4 @@
 export default defineNuxtConfig({
   extends: '@nuxt-themes/docus',
+  modules: ['@nuxtjs/plausible'],
 })