+++ docs/tokens.config.ts
@@ -1,7 +1,19 @@
-import { defineTheme, palette } from 'pinceau'
+import { defineTheme } from 'pinceau'
 
 export default defineTheme({
   color: {
-    primary: palette('orange'),
+    // https://www.tints.dev/primary/14CFBD
+    primary: {
+      50: '#E3FCFA',
+      100: '#CCFAF5',
+      200: '#99F5EC',
+      300: '#66F0E2',
+      400: '#32EBD9',
+      500: '#14CFBD',
+      600: '#10A798',
+      700: '#0C7E72',
+      800: '#08544C',
+      900: '#042A26',
+    },
   },
 })