+++ packages/devtools-ui-kit/src/runtime/client.ts
@@ -1,0 +1,3 @@
+import { useDevtoolsClient } from '@nuxt/devtools/iframe-client'
+
+export const devToolsClient = useDevtoolsClient()