+++ packages/devtools/client/components/ComponentsList.vue
@@ -4,8 +4,9 @@ import Fuse from 'fuse.js'
 
 const props = defineProps<{
   components: Component[]
-  search: string
 }>()
+
+const search = ref('')
 
 const fuse = computed(() => new Fuse(props.components, {
   keys: [
@@ -28,18 +29,12 @@ const filtered = computed(() => {
     runtime: 0,
   }
 
-  const result = props.search
-    ? fuse.value.search(props.search).map(i => i.item)
+  const result = search.value
+    ? fuse.value.search(search.value).map(i => i.item)
     : props.components
 
   result
     .forEach((component) => {
-      if (component.global) {
-        runtime.push(component)
-        count.runtime++
-        return
-      }
-
       if (component.filePath && isNodeModulePath(component.filePath)) {
         const name = getModuleNameFromPath(component.filePath)
         if (!name)
@@ -54,6 +49,10 @@ const filtered = computed(() => {
           lib.get(name)!.push(component)
           count.lib++
         }
+      }
+      else if (component.global && !component.filePath) {
+        runtime.push(component)
+        count.runtime++
       }
       else {
         user.push(component)
@@ -73,15 +72,26 @@ const filtered = computed(() => {
 </script>
 
 <template>
-  <SectionBlock
+  <div p4 flex="~ gap4" flex-1 border="b base" navbar-glass>
+    <NTextInput
+      v-model="search"
+      placeholder="Search..."
+      icon="carbon-search"
+      p="x5 y2"
+      n="primary"
+      flex-auto bg-base border-base
+    />
+    <slot />
+  </div>
+  <NSectionBlock
     v-if="filtered.user.length"
     icon="carbon-nominal"
     text="User components"
     :description="`Total components: ${filtered.count.user}`"
   >
     <ComponentItem v-for="c of filtered.user" :key="c.filePath" ml--5 :component="c" />
-  </SectionBlock>
-  <SectionBlock
+  </NSectionBlock>
+  <NSectionBlock
     v-if="filtered.runtime.length"
     icon="carbon-load-balancer-global"
 
@@ -89,26 +99,26 @@ const filtered = computed(() => {
     :description="`Total components: ${filtered.count.runtime}`"
   >
     <ComponentItem v-for="c of filtered.runtime" :key="c.filePath" ml--5 :component="c" />
-  </SectionBlock>
-  <SectionBlock
+  </NSectionBlock>
+  <NSectionBlock
     v-if="filtered.builtin.length"
     icon="simple-icons-nuxtdotjs"
     text="Built-in components"
     :description="`Total components: ${filtered.count.builtin}`"
   >
     <ComponentItem v-for="c of filtered.builtin" :key="c.filePath" ml--5 :component="c" />
-  </SectionBlock>
-  <SectionBlock
+  </NSectionBlock>
+  <NSectionBlock
     v-if="filtered.lib.size"
     icon="carbon-3d-mpr-toggle"
     text="Components from libraries"
     :description="`${filtered.count.lib} components from ${filtered.lib.size} packages`"
   >
     <div v-for="[key, value] of filtered.lib.entries()" :key="key">
-      <IconTitle :text="`${key} (${value.length})`" op50 py1 />
+      <NIconTitle :text="`${key} (${value.length})`" op50 py1 />
       <div pl4>
         <ComponentItem v-for="c of value" :key="c.filePath" :component="c" />
       </div>
     </div>
-  </SectionBlock>
+  </NSectionBlock>
 </template>