+++ packages/devtools/client/composables/client-services/shiki.ts
@@ -13,10 +13,11 @@ getHighlighter({
     'css',
     'javascript',
     'typescript',
+    'html',
   ],
 }).then((i) => { shiki.value = i })
 
-export function highlight(code: string, lang: Lang) {
+export function renderCodeHighlight(code: string, lang: Lang) {
   const mode = useColorMode()
   if (!shiki.value)
     return code