+++ packages/devtools/client/pages/modules/components.vue
@@ -1,5 +1,7 @@
 <script setup lang="ts">
 import { useComponents } from '~/composables/state'
+import ComponentsGraph from '~/components/ComponentsGraph.vue'
+import ComponentsList from '~/components/ComponentsList.vue'
 
 definePageMeta({
   icon: 'carbon-nominal',
@@ -10,7 +12,6 @@ definePageMeta({
 const client = useClient()
 const router = useRouter()
 const components = useComponents()
-const search = ref('')
 
 const {
   componentsView: view,
@@ -29,32 +30,27 @@ function toggleView() {
 </script>
 
 <template>
-  <div h-full of-auto>
-    <div p4 flex="~ gap4" flex-1 border="b base" navbar-glass>
-      <NTextInput
-        v-model="search"
-        placeholder="Search..."
-        icon="carbon-search"
-        p="x5 y2"
-        n="primary"
-        flex-auto bg-base border-base
-      />
-      <button
-        title="Toggle view"
-        @click="toggleView"
-      >
-        <NIcon v-if="view === 'graph'" icon="i-carbon-network-4" />
-        <NIcon v-else icon="i-carbon-list" />
-      </button>
-      <button
-        v-if="client?.inspector?.instance"
-        title="Inspect Vue components"
-        @click="openComponentInspector"
-      >
-        <NIcon icon="i-carbon-select-window" />
-      </button>
-    </div>
-    <ComponentsGraph v-if="view === 'graph'" :components="components" />
-    <ComponentsList v-else :search="search" :components="components" />
+  <div h-full of-auto relative>
+    <component
+      :is="view === 'list' ? ComponentsList : ComponentsGraph"
+      :components="components"
+    >
+      <div flex-none flex="~ gap4">
+        <button
+          title="Toggle view"
+          @click="toggleView"
+        >
+          <NIcon v-if="view === 'graph'" icon="i-carbon-list" />
+          <NIcon v-else icon="i-carbon-network-4" />
+        </button>
+        <button
+          v-if="client?.inspector?.instance"
+          title="Inspect Vue components"
+          @click="openComponentInspector"
+        >
+          <NIcon icon="i-carbon-select-window" />
+        </button>
+      </div>
+    </component>
   </div>
 </template>