+++ packages/devtools/client/pages/modules/imports.vue
@@ -86,29 +86,29 @@ const filtered = computed(() => {
         </NSwitch>
       </div>
     </div>
-    <SectionBlock
+    <NSectionBlock
       v-if="filtered.user.size"
       icon="carbon-function"
       text="User composables"
       :description="`${filtered.count.user} composables from ${filtered.user.size} modules`"
     >
       <ComposableTree :map="filtered.user" :root="config.rootDir" :metadata="importsMetadata" />
-    </SectionBlock>
-    <SectionBlock
+    </NSectionBlock>
+    <NSectionBlock
       v-if="filtered.builtin.size"
       icon="simple-icons-nuxtdotjs"
       text="Built-in composables"
       :description="`${filtered.count.builtin} composables`"
     >
       <ComposableTree :map="filtered.builtin" :root="config.rootDir" :metadata="importsMetadata" />
-    </SectionBlock>
-    <SectionBlock
+    </NSectionBlock>
+    <NSectionBlock
       v-if="filtered.lib.size"
       icon="carbon-3d-mpr-toggle"
       text="Composables from libraries"
       :description="`${filtered.count.lib} composables from ${filtered.lib.size} packages`"
     >
       <ComposableTree :map="filtered.lib" :root="config.rootDir" :metadata="importsMetadata" />
-    </SectionBlock>
+    </NSectionBlock>
   </div>
 </template>