+++ packages/devtools/client/pages/modules/virtual-files.vue
@@ -1,4 +1,5 @@
 <script setup lang="ts">
+import { Pane, Splitpanes } from 'splitpanes'
 import Fuse from 'fuse.js'
 
 definePageMeta({
@@ -72,8 +73,8 @@ const filteredFiles = computed(() => {
 </script>
 
 <template>
-  <div grid="~ cols-[auto_1fr]" h-full of-hidden class="virtual-files">
-    <div border="r base" of-auto resize-x w="300px">
+  <Splitpanes h-full of-hidden class="virtual-files">
+    <Pane border="r base" size="30" min-size="5" of-auto>
       <div pb2 p3>
         <NTextInput
           v-model="searchString"
@@ -90,15 +91,17 @@ const filteredFiles = computed(() => {
       >
         {{ toShortPath(f.id) }}
       </NuxtLink>
-    </div>
-    <div v-if="current?.content" h-full of-hidden flex="~ col">
-      <div border="b base" px4 py2 text-sm op75 flex-none>
-        <code>{{ current.id }}</code>
+    </Pane>
+    <Pane min-size="5">
+      <div v-if="current?.content" h-full of-hidden flex="~ col">
+        <div border="b base" px4 py2 text-sm op75 flex-none>
+          <code>{{ current.id }}</code>
+        </div>
+        <NCodeBlock of-auto h-full text-sm :code="current.content" lang="typescript" />
       </div>
-      <pre of-auto h-full text-sm v-html="highlight(current?.content, 'typescript')" />
-    </div>
-    <span v-else flex items-center justify-center op50>Select one file to start</span>
-  </div>
+      <span v-else flex items-center justify-center op50 h-full>Select one file to start</span>
+    </Pane>
+  </Splitpanes>
 </template>
 
 <style>